requirements.txt
aiosmtplib==3.0.2
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
bcrypt==4.0.0
certifi==2024.12.14
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.0
dill==0.3.9
distro==1.9.0
docopt==0.6.2
ecdsa==0.19.0
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inquirerpy==0.3.4
Jinja2==3.1.5
jiter==0.8.2
Mako==1.3.8
MarkupSafe==2.1.5
multiprocess==0.70.14
nestd==0.3.1
openai==1.58.1
orjson==3.9.15
passlib==1.7.4
pfzy==0.3.4
prompt_toolkit==3.0.48
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.2.1
robyn==0.64.2
rsa==4.9
rustimport==1.3.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stomp.py==8.2.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
watchdog==4.0.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
zipp==3.20.2
